requirements.txt
astroid
autopep8
Django>=2.1.11
django-crispy-forms
django-likeit
djangorestframework
drf-nested-routers
django-extensions
isort
lazy-object-proxy
mccabe
Pillow
pycodestyle
pylint
pytz
six
virtualenv
wrapt
django-heroku
psycopg2-binary
django-cors-headers
djangorestframework-jwt
djangorestframework-jsonapi
Faker
gunicorn